Optimization of multiple functions, replace old opt
